+++ src/kerberos/datagram.rs
@@ -33,7 +33,7 @@ impl DatagramParser for KerberosUdpParser {
         _ts: Timestamp,
         out: &mut Vec<Self::Message>,
     ) {
-        if let Some(msg) = parse(payload) {
+        if let Ok(msg) = parse(payload) {
             out.push(msg);
         }
     }